[feature/#33-conversational-interface] 언어 기반 분기 함수 분리
