+++ src/hooks/useLocalRecognition.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { LANG } from '@constants/setting';
 import { ServiceType, SystemStatus } from '~/types/common';
 import { checkPause } from '~/utils/audio';
-import { handleError } from '~/utils/common';
+import { handleError } from '@utils/error';
 import {
 	setDialogueStateBypassPause,
 	useAppDispatch,
